Renamed "rules" package to "configuration" and created HeuristicWeights for future purpose
